requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
bcrypt==4.1.2